Show error messages in line with address select fields (#225)
* Show error messages in line with address select fields

While working on setting up spec tests for the errors we show to the user, we found that the address lookup pages were the only pages not showing the message in line with the field.
Seems like some of the code necessary for the field to show was there but pointing to `addresses` rather than `temp_addresses`. We have confirmed that the `addresses` field is not any more used and so we have removed it and we are checking the `temp_address` field only instead.