Create a new transient registration for WEX renewals (#236)
https://eaflood.atlassian.net/browse/RUBY-491

When a user starts the WEX renewal process, we need to create a new database record to store their in-progress renewal.

As a start, we need to create a new RenewingRegistration subtype of TransientRegistration.

We also create the RenewalStartService. Currently this just creates a new RenewingRegistration record in the database.

Finally we call RenewalStartService from a new route. This is a temporary route for the purposes of testing and meeting the acceptance criteria. We output the JSON from the new record, also for testing purposes.