Change position of the Exemptions form (#239)
* Change position of the Exemptions form

From: https://eaflood.atlassian.net/browse/RUBY-457

Change the state machine and fix tests so that the exemptions screen will now appear just after the location screens.

Other fixes were necessary on the test suite for all this to run. The first thing that went wrong with them, is that removing the exemptions form between the site address and check your answers screen was generating an issue with the helper we use to fetch the correct `previous state` of a form. The code that was there was doing very little on fetching the correct previous status, was not accounting for multiple options, like is the case when dealing with the site_address, and was using "next" AASM events rather than "back" ones, so it was actually working just by chances rather than correct logic.
Once that was fixed to return the correct previous state, it was now breaking the logic for "flexibly navigatable" tests in the "get form" share examples. This is because we rely on the get_previous_state to return a result for forms that should not have any previous state, like is the case for not flexibly navigatable forms, hence the fixes there.