Complete renewal and create new registration (#247)
https://eaflood.atlassian.net/browse/RUBY-517

When a user reaches the "renewal complete" page, we should create a new registration with the data they just confirmed. The renewing_registration record should also be deleted from the database.

This is done by calling RegistrationCompletionService, the same way as completing a brand new registration does.

We also display the reference number of the newly-completed registration. This will be styled properly as part of RUBY-500, but in the meantime it's useful for debugging.

This commit also includes light refactoring of RegistrationCompletionService to inherit from BaseService. This means we can just call `RegistrationCompletionService.run` instead of having to initialise and then make a second call to do things. It also brings our testing more in line with the other services.