Make it possible to use ENV variables in feature toggles
In order to use ENV variables we are using ERB to parse the YAML file before feeding the result to the YAML library.
To test it, another fix was necessary on the test suite, as Dotenv needs extra rules to load the `.env` from the dummy app placed in the `spec` directory.