Add options to renewal start page (#256)
https://eaflood.atlassian.net/browse/RUBY-493
https://eaflood.atlassian.net/browse/RUBY-498

These buttons will be used to route the user to the declaration or through the longer journey through the forms.

This means we need to add temp_renew_without_changes to DB. We'll need to store this value so we can check the user's choices again later, or replay it if they return to the page. This follows the model we use for other radio buttons that are just used for routing, like the tier check skip in waste carriers.