Set up 'Do you want to renew with these details?' page (#255)
https://eaflood.atlassian.net/browse/RUBY-493

Users should reach this page after clicking the magic link.

To save time, we'll base this on the existing "check your answers" page. However, it does have a different header and different options at the bottom (“Renew with these changes”/”I need to make changes”).

In summary, this means refactoring to move the bits we want to share into partials, and then plopping them on the RenewalStartForm and rearranging as needed.

* Move 'check your answers' sections to partials

This will let us re-use them on the renewal journey, but in a different order if needed.

* Move exemptions to the top of the page

This is now at the start of the journey, so should be at the start of the content for the page.

* List partners if there are any

* Move attribute assignment to a separate method

This will make us easier to move the code that we'll share with the renewal start form into a module

* Create DataOverviewForm module

This follows the pattern we've set with AddressForm. This should include all the data-setting and displaying bits that we want to share between CheckYourAnswersForm and RenewalStartForm.

* Assign and display attributes on RenewalStartForm

Now we take all of the "check your answers" bits we moved into shared code and slap them on this form.

The only difference is that we move the 'site' section up to just under exemptions to match the wireframe.

* Fix form option spacing

* Move applicant details to match wireframe

* Change headings to match wireframe