Use custom message for validation error
While testing this, I spotted a bug in the validator gem where we are assigning the error to the wrong field: https://github.com/DEFRA/defra-ruby-validators/pull/34

The result is that the error message does not appear a second time directly over the question (it also would break the link from the top error, but that has no impact at the moment as it's broken anyway).

Once the gem is updated, that should be fixed.