Fix bug on renewal creation (#271)
Two fixes were necessary:
The first is about adding a `&` on the Airbrake error reporting code for when an error is thrown before the `@registration` object is initialized.
The second fix is the actual problem. I was using the `token` from the Transient Registration to populate a referring registration, but that is incorrect as the 2 tokens are not going to match. What is going to match is instead the `reference`, and hence I have changed the code to use that in order to fetch the correct referring registration.