Update W3C support to deal with multiple scenarios on same test
This little change will create a folder per test in our cassettes directory so that we can support multiple scenarios on the same test having to check different pages or different version of the same page. This is because currently the support system only wants one registration per file and will also reduce maintainability since we won't need to regenerate an existing cassette when adding a scenario.