Don't use default scope to order registration exemptions (#275)
https://eaflood.atlassian.net/browse/RUBY-520

Our first pass at ordering registration exemptions used a default scope. However, this broke some scopes in the host apps (see https://github.com/DEFRA/waste-exemptions-back-office-ta/pull/323) and when looking into it, all the advice was "don't use default scopes".

So this change makes it a non-default scope and then applies it to the locations we want to order.