Remove unnecessary data after business type change
https://eaflood.atlassian.net/browse/RUBY-503

Some data is only required for registrations with certain business types - partnerships need partners, while limited companies and LLPs need Companies House numbers.

However, if the business type changes from one which requires this data to one which doesn't, we should remove the extra data.

This could be because the user changes the business type during renewal, but it could also happen during a new registration if they go through most of the process and then go back to change their business type.