Capitalise env var in feature toggle test (#283)
This test was failing for me on Vagrant, apparently because of a difference in the cases (we add a lowercase env var, but the feature toggle expects an uppercase one). It looks like it passes elsewhere so it may depend on the operating system or something, but seemed work a fix.