Set up custom dimensions to track workflows
https://eaflood.atlassian.net/browse/RUBY-509

We want to be able to track the performance of the new registration journey and the renewals journey. Because they use the same pages, we'll need a different way to differentiate.

This PR implements custom dimensions in the data layer to be used by Google Tag Manager and Google Analytics. This will allow us to filter pageviews by the journey, and will potentially let us do even more with the data layer in the future.