Display error when trying to renew too late (#274)
https://eaflood.atlassian.net/browse/RUBY-496

We have certain business rules for when a registration can and cannot be renewed.

This includes a 'grace period' after the registration expires when it can still be renewed. This is 30 days after the expiry date.

If an expired registration is no longer in the grace period, it should not be possible to renew it. Anyone trying to access the renewal at any point in its process (clicking the magic link, loading a page midway through the renewal process) should see a useful page explaining that the registration can no longer be renewed.

As part of this change, we renamed the env vars used for the renewal window and moved them to the engine initialiser. This will necessitate some changes in the apps for where we load these env vars (in the engine initialiser rather than application.rb).

We also moved the renewal window methods into the engine. These methods existed in the back office, but we have agreed to move them into the engine as they will now be needed in there as well.