Fix past_renewal_window? to not drop one day
It was spotted in testing that we were only actually providing a grace period of 29 days. Hopefully the following scenario helps explain the situation.

- We have a registration that expires on 31 Jul 2019 (24:00)
- 31 Jul 2018 plus 30 days is Aug 30
- If the current time is 14:34 on Aug 30 `past_renewal_window?` would return true
- it should not return true until 00:00 on 31 Aug

The logic prior to this change was

```ruby
def past_renewal_window?
  (expires_on + renewal_window_after_expiry_in_days.days) < Time.now
end
```

Initially we thought we could fix it by changing the comparison to `<=` however that still leads to the comparison returning `true`. The issue is actually the use of `Time` in a comparison that uses just date objects. Using this time and our example above you get `00:00 30 Aug 2019 <= 14:34 30 Aug 2019`.

Hence this change which updates the comparison to be `< Date.current`. Using the example above means the test becomes `30 Aug 2019 < 30 Aug 2019`.