Only display type-dependant data when required
https://eaflood.atlassian.net/browse/RUBY-503

We recently made a change to remove type-dependant attributes from a renewed registration when a business type change meant they were no longer needed.

However, as this change happens in the RegistrationCompletionService, looking at 'check your answers' still displays the now-unneeded data, which the user may now be unable to modify.

To avoid confusion, this PR hides this data when it is no longer in use and will not be saved.