Check for HTML-escaped strings in request tests
We recently added a number of request tests for the edit and renewal journeys. These tested to make sure that values assigned to the registration were pre-filled in the forms.

However, these tests failed if the string value included characters which were then HTML-escaped - for example, `Hyphen-O'Postraphe` became `Hyphen-O&#39;Postraphe`, and the match failed. This led to a lot of intermittent failures.

This PR makes sure we also escape the value before trying to match it, which should stop these specs from failing.