Copy expired registrations to RenewingRegistration
https://eaflood.atlassian.net/browse/RUBY-594

When we start a renewal, if the registration has already expired and is in the grace period, we want to copy over the expired exemptions so they don't have to re-select them.

This will also mean the exemptions are listed in the magic link email if it is resent after expiry.