Fix existing data copying tests
We weren't testing with a completed registration, so there weren't any addresses or people to copy. This meant that we were testing that we had successfully copied 0 things from 0 things, which is pointless.

Testing with a complete registration showed the address tests didn't work, so this also fixes the tests.