Copy expired registrations to RenewingRegistration (#303)
https://eaflood.atlassian.net/browse/RUBY-594

When we start a renewal, if the registration has already expired and is in the grace period, we want to copy over the expired exemptions so they don't have to re-select them. (But for editing, we only want to copy active ones.)

While working on this issue, I also fixed the existing data-copying tests. We weren't testing with a completed registration, so there weren't any addresses or people to copy. This meant that we were testing that we had successfully copied 0 things from 0 things, which is pointless. Testing with a complete registration showed the address tests didn't work, so this PR includes fixes to the tests.