Redirect to start form page when a token is invalid
Part of: https://eaflood.atlassian.net/browse/RUBY-603

When a user hits a form page with an invalid token, the current code will generate a `NewRegistration`, assign it the workflow state for the form the user has requested and try to render the page. The problem is that if a registration is created in this way, whithout ever going through the first `start_form` submit, that registration has no token assigned to it. Hence, when the view renders, it will generate an error as the back button url cannot be generated.

This change will make sure that when a token is invalid and hence a `NewRegistration` is created and not persisted, its status is not changed at all from the default `start_form` and so the user is redirected to the start form page.