Add `#renewal?` method to registration objects
Part of: https://eaflood.atlassian.net/browse/RUBY-506

This adds a common interface method on the registration classes.
A `Registration` is a `renewal` if it refers to another registration that exists.
A `TransientRegistration` is by default *not* a renewal, while a `RenewingRegistration` is by default a renwal.

This code is added to the engine in order to be reused in the backoffice