Make confirmed email an attribute of the form
As part of the form refactoring, we want to be able to use delegate on our attributes and not re-write main initialize and submit methods.
In the case of the applicant confirmed email, we want the form to deal with the attribute in this different way, were we are able to set and read it and overwrite its value.
When sorting out the `submit` method, we will also change the setting of the attribute from params making use of callbacks with parameters.