Forms refactoring - Exemptions form division of responsibilities
Leave to the view what belongs to the view and to the form what belongs to the form :D

This change remove the list of exemptions being assigned to the form using the attribute name that is suppose to return the transient registration exemptions. It instead uses it as the real list of exemptions that are part of a registration or assign to them the registration ticked in the form itself. An helper is instead responsible to fetch the list of all exemptions for the view to show.