Address form - Remove AddressForm and create an AddressHelper
As part of the form refactoring, I believe that the AddressForm module is not needed. The only actually implemented method is by its nature more an helper than an object method. So In order to reduce complexity of these classes, a first step for me was to move the add_or_replace_address method into an AddressHelper class. It might get moved again or deleted if we find a way to achieve it differently after we complete the form refactoring