Use rails associations to define associations with the addresses
In both the Registration and Transient registration we are implementing Rails association in order to associate those registration with the correct addresses.
We have decided that the name of the methods should be the same for bot Transient registration and non transient registration.
This allow for us to be able to easely reuse the same code on the 2 kind of object, and at the same time gives us the ability to edit and update the addresses on those registration without having to deal with the addresses array association complication.