Change urls resource to start with token
We have multiple little issues due to the fact that `POST :create` requet expect the `:token` param from the attributes rather than the url.
Issues includes, but are not limited at, the browser back button redirecting back to a 500 error after a form have been posted in an invalid format.

In order to make sure that our `:create` requests always get a token through the `url` as per our `new` requests, we have moved the token away from the single resources requests, to always appear before the resource paths.

On the code itself, minor changes were needed, where particular forms were expecting the token to appear as part of an object attribute. Other fixes are about the way we now compose the url using helpers.
All other fixes are to do with the test suite paths.