Remove exemptions checkbox from inside label
From: https://eaflood.atlassian.net/browse/RUBY-608

We have tried to restyle the GDS elements checkbox to try achieve better reading for screen readers on our exemptions list.
In doing so, however, we have created a visual issue in the fact that the `focus` status's stile does not get applied to the checkboxes.
We have then decided to wait for GDS to deal with the screen readers issue (a ticket have been opened with them and is in their backlog).
So we are rolling back a part of this refactoring and the custom style in order to restore the checkboxes to their original behaviour.