Report failed response to Errbit
If the lookup fails when call the WFS we want to report the error with our Errbit service so we can monitor the reliability of the calls we're making, and see if the service has gone down.

As part of this came across the fact that though Airbrake is a key dependency of the engine, we don't specify it in the gemspec!

For some reason this gem we've just relied on the host apps adding it to their Gemfiles.

We need it here to allow us to mock Airbrake in our tests, but once this PR is merged we'll look to remove Airbrake as a direct dependency in the front and back office apps.