Set the area when adding a site address
This adds the final piece of functionality which is to update the area in the data hash which is generated and used to create the transient site address.

It includes some refactoring to reduce some duplication that was appearing in the methods to handle the extra functionality site addresses need.