Identify EA Administrative Area for new regs. (#347)
The previous WEX service would attempt to identify the EA Admin area for a site once it had been entered as part of the registration. This stopped working in December 2018.

When we shipped the new version of the service we left out this functionality because

- we were still awaiting a fix
- we knew both FRAE and WEX depend on this so wanted to implement a shared solution

With the creation of [defra-ruby-area](https://github.com/DEFRA/defra-ruby-area) we now have that shared solution.

So this change covers everything needed to integrate the gem and identify the area as part of the new registration journey.

N.B. Whilst doing this we came across the fact that though Airbrake is a key dependency of the engine, we don't specify it in the gemspec!

For some reason this gem just relied on the host apps adding it to their Gemfiles. Our intention is with this merged we can remove the explicit references to Airbrake in the host apps Gemfile's.