Refactor Contact Address Manual form (#341)
* Refactor Contact Address Manual form

This removes inheritance just so the class is in a clear state and so I can work on the different address form separately. Once the refactor of the BaseForm is done, will be easier to put together common bits of code (With CodeClimates help too!).
The parameters, in this case, are coming through the StrongParameter Rails implementation.
This allow for `accept_nested_attributes` to work, which will then allow us to implement the `delegate` approach to the forms.
The object passed to the shared partial is also changed, so that we receive the parameters correctly nested.
Everything else is fixes around that.
Particularly the specs were to change so that they will pass over proper params object. This do not affect existing forms still dealing with hashes, as the params object does behave as such already.
We already agreed that strong parameters will be the way to go, and once that's completed then we can move the implementation away from the forms and back into the controller.