Revert "Restrict address missing scopes to site addresses (#365)" (#366)
This reverts commit 5556ffacd07d339af54d34deb28c347ac2810683.

The change made was due to a lack of understanding about scopes, and the fact that because we use enums that we get the scope `:site` for free on the address.

So rather than changing the current scopes to include a condition to test for the address type, instead we can just chain scopes together

```ruby
WasteExemptionsEngine::Address.site.missing_area
```

This gives the same result as `sites_missing_area`. By reverting this change our scopes become flexible again, and we can do things in a much more 'rails' way.