Site Address manual form refactoring
Following on: https://github.com/DEFRA/waste-exemptions-engine/pull/341/files
This apply the same refactoing to the site address manual form.
The goal is to make use of rails-ness on these forms, so we want them to name the attributes as per TransientAddress attributes. This will allow the use of strong parameters and less code to custom-submit these forms rather than letting rails do the job for us :D