Add with_postcode scope to Address
https://eaflood.atlassian.net/browse/RUBY-486

Will resolve an issue in x & y automated lookup in the [back-office](https://github.com/DEFRA/waste-exemptions-back-office) which is currently trying to find the x & y values for addresses with no postcode.

There are not many, but there are some addresses in the database where we literally have no current information, including a postcode. The x & y job needs a postcode to perform the lookup on OS Places, so we need a scope to allow us to filter these records out to prevent errors.