Update Address scope tests to use match_array
This is just some housekeeping. It was spotted that we were not being consistent in our tests for the scopes on the Address model in a recent commit to add a `with_postcode` scope.

So this change updates all the tests to use `match_array` rather than having separate expectations to check the size and content of the returned result.