Set version of defra_ruby_validators
https://github.com/DEFRA/waste-exemptions-front-office/pull/256
https://github.com/DEFRA/waste-exemptions-back-office/pull/435

We just had Dependabot update the version of [defra_ruby_validators](https://github.com/DEFRA/defra-ruby-validators) used by the front and back office under the guise of updating the engine.

This broke the apps, so to protect against this in future we are using pessimistic versioning for the gem to ensure the apps don't try and take a version they're not ready for.