Fix problem with area update for address lookup
The problem with the area not being populated for site addresses created from an address lookup was that the refactoring started passing in a symbol rather than the expected enum.

This change keeps the refactoring, but corrects what we pass in to `TransientAddress.create_from_address_finder_data` so its a recognised enum.