Refactor the site address trait in Factories (#388)
Ahead of the need to add some tests specific to site addresses we spotted that the current `:site_address` trait sets a couple of other fields, not just the address type. This is different to the `:operator_address` and ':contact_address` traits which just set the address type

We need to set the address type as part of some tests we are looking to add. But we also need to be able to specify creating the 3 different kinds of site address; from grid reference, from address lookup, and from address manual.

In its current state we would not have been able to reuse the trait. With this refactor we can, plus

- its now consistent with the other address type traits
- the new trait is more descriptive of what is actually going on