Remove token from forms
Following the bug fix on the back button issue which brought us to the change in the position of the token in the url, now we can assume Tokens are always going to be available through every form request, and hence we can get rid of the token instnces in the forms views.
The only exception will continue to be the `start` page, which should not need a token to work but that will need to work with a token too. This avoids the user going back too many times, reaching the start page and having to re-do the registration from scratch because the token is not available. Hence, in the `get` request the token is currently a main parameter in the URL request, while for `post` requests it will be send throught as a main hidden-field-tag attribute.
 This means that in the controller it will always be accessible in the same way, using `params[:token]`