Update x & y from manual postcode
Add into the call back functionality to update the x & y for site addresses from the postcode, if the site was entered using a manual address.

This has admittedly got a little messy with how we interact with the AddressFinderService, because of the way it returns its results.

Hopefully this covers all eventualities though!