Refactor hooks to have 1 site check & 1 save call
We've refactored the calls to multiple after_create methods into a call to one. Now we only perform the `site?` check and call to `save!` to persist the changes once.

The downside is we have complexity issues with rubocop. I've chosen to add an exception here as I think what is currently there makes sense, and the reasons provided are valid.

However I'm open to suggestions. If there is a way we can reduce the calls to the db and the duplication whilst still solving the complexity issue properly I would love to hear it!