Fix broken renewal test
be rspec ./spec/models/waste_exemptions_engine/renewing_registration_spec.rb:53

Now we have an `after_create` hook in the `TransientAddress` this test was failing because it would create an `Address` without a populated x & y using FactoryBot, then create the renewing registration which in turn creates transient addresses.

The hook was then firing on these leading the test to fail because the attributes of the original registration and its site address did not include an x & y value.

To better reflect the state we would expect to find an `Address` record in, we added the x & y values to the trait.

N.B. This broke other tests as well but fixing this one seems to have fixed the others affected by this as well.