Refactor service to return a result instead
As a means of making the service more testable, this changes the service's behaviour to return the x and y rather than populating it.

It means in our `TransientAddress` tests where the intent is to test the model and not this service, we can mock it using `expect(XAndYFromAddressService).to receive(run).with(subject).and_return(x: 123, y: 321)`.

We can then move those tests that are actually confirming the behaviour of the service to a new spec, and not end up duplicating anything.