Add determine grid reference service
Following on from the service to determine the x and y values based on an existing grid reference or postcode, this adds a new service that determines a grid reference based on x & y (easting and northing) values.

This continues the move to simplify the `TransientAddress` and move some of the existing logic into Single Responsiblity classes.