Add new valid_x_and_y? method to TransientAddress
Realised our current checks in TransientAddress to see if we should call the new services were not complete enough.

For example we were checking whether x & y were blank before calling AreaLookupService, but not whether they were set to 0.0. If this was the case we'd just generate an error and no result for no reason.

So added a new method to determine whether the x and y are values are valid and can be used to determine other details.