Update transient address and spec
We've updated the update site detail methods to make use of the new `valid_x_and_y?` method.

We have also updated the specs. They are still essentially checking for the same end result, however we've combined some to reduce the time and load, and stubbed all the underlying services. If we didn't there would be little point in having unit tests directly for them, as they would be tested through these.

However unit tests are also a way of documenting the expected behaviour of a class, and helping in its design. So we want unit tests for the underlying services, but we don't want to duplicate them. So we have moved to stubbing them in these tests.