Rename area lookup service to be consistent
We now have a service that determines grid references based on coordinates, and one that determines x & y values based on coordinates. As these use the prefix `Determine` it makes sense to rename the `AddressLookupService` to be consistent with them.

N.B. This will break the back office but that's a quick change to resolve.