Add checks to ensure external service is called correctly
Was spotted in review that there is a typo in the `DetermineAreaLookup` logic that is not being covered by our tests (they are all still passing).

We've avoided using VCR for the area lookup because the WFS is so flaky. For example if travis has to fall back to actually calling the service there is a good chance it will fail because it is down or not responding quickly enough. We'll get a test fail that telling travis to rebuild will fix.

So to avoid the flakiness with have stubbed it out completely. This change though ensures we are at least calling it in the way it expects.