Refactor area, grid reference and x & y auto population (#384)
https://eaflood.atlassian.net/browse/RUBY-483

It was spotted during testing of the area lookup integration that some of the extra functions that happen when we add an address had stopped working.

After some investigation it looks like the work we have been doing to refactor and simplify the forms used for addresses has broken a few things 😢

This change fixes the issue, but a big factor in why it broke was the complexity in the way we currently had those extra functions implemented. So this change also takes the opportunity to refactor this area of the code base.

In essence when a site address is added 3 extra bits of information need to be determined

- the grid reference (if an address rather than grid reference is provided)
- the x & y coordinates (if the address lookup wasn't used)
- the EA administrative area

We've simplified the code, particularly in the TransientAddress by moving these functions into their own services, which are then manged by a fourth.

Also rather than using class methods on the TransientAddress for creating addresses along with this extra information, we now use a `before_create` hook to fire off these extra services and determine the information. So the logic for creating a manual address has been removed as its no longer used, and create_from grid and address lookup will going soon as well.