Update address base url details (#399)
* Update address base url details

In our local environment we hadn't quite got the configuration of the address base facade correct.

Turns out this was blocking vagrant from using the cassettes when running tests locally, because it could no longer match the requests.

This change fixes the issue.

## Note

When attempting to run the test suite locally rspec complained there were migrations pending in the dummy app.

A `RAILS_ENV=test be rake db:` `drop`, `create`, and `migrate` led to the schema being updated. After a discussion with @cintamani it looks like we manually updated the schema to match something in the back office because of a miscommunication.

Hence running the migration led to something being dropped rather than added from the schema.